requirements.txt
pillow
torch
realesrgan
